added a person server that subscribes to a message queue.
Currently no send, or knowledge of other queues. An idea on this is maybe a 'person' registry, that tracks all people and you can lookup, can send a message to a person as this would know about person + their associated queue.
Need to update a message to a know about who sent it.